Add extended characters handling to DeuLigne
This commit creates a mapping (in Deuligne.h) to display extended characters using the 8 "CGRAM" chars rather than from the CGROM mapping.
This way, there is no dependancy on the actual ROM mapping in the display - chars are stored on the fly in one of the 8 CGRAM position.

Issues:
* Only 8 extended chars are loadable at any given time. If a 9th need to be stored, then one of the least used one is evicted.
* The extended display of the character must be transfered to the display , this takes time.

Still need to be done:

* Before evicting a character, check that it's not currently displayed (by reading the data). Will slow this particular print,
  but shall not happen often.

* Store character table in progmem